+++ src/metrics/monitor/getUptime.ts
@@ -21,22 +21,19 @@
  * SOFTWARE.
  */
 
-/// <reference types="node"/>
+import { createMetric } from "./builder/create";
 
-import * as http from "http";
+const DefaultFormatter = (value: number) => "" + value;
 
-import { createMetricsAsync, GetMetricsResolve } from "./metrics/createMetricsAsync";
-
-export async function getMetricsAsync() {
-    return createMetricsAsync();
-}
-
-export async function listen(port = 9000) {
-    http.createServer(async (request: http.ServerRequest, response: http.ServerResponse) => {
-        const metrics = await createMetricsAsync();
-
-        response.writeHead(200, { "Content-Type": "text/plain" });
-        response.write(metrics);
-        response.end();
-    });
+/**
+ * Creates a metric from uptime process.
+ *
+ * @param {Process}     [currentProcess]    A current process.
+ * @param {string}      [name]              A metric name.
+ * @param {Function}    [formatter]         A formatter.
+ * @return {string} The metric.
+ */
+export function getUptime(currentProcess = process, name = "node_process_uptime", formatter = DefaultFormatter) {
+    const uptime = formatter(currentProcess.uptime());
+    return createMetric(name, uptime);
 }