+++ CHANGELOG.md
@@ -2,6 +2,17 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [0.2.0] - 2017-03-24
+
+### Added
+- Add a `listen` function to create a small web service;
+
+### Fixed
+- Fix missing TypeScript definition;
+
+### Removed
+- Remove options;
+
 ## [0.1.3] - 2017-03-18
 
 ### Changed
@@ -18,4 +29,5 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ### Fixed
 - Fix TypeScript error (private formatter `Function`);
 
+[0.2.0]: https://github.com/cyrilschumacher/prometheus-node-usage/compare/0.1.3...0.2.0
 [0.1.3]: https://github.com/cyrilschumacher/prometheus-node-usage/compare/0.1.2...0.1.3